adjust functions
